Attempt to fix s3 synapse bucket provisioning
Seeing the following error when provisioning an S3 synapse bucket:

"errorCode":"ValidationException","errorMessage":"Processed template of
Stack SC-237179673806-pp-pdfndguyqnu3k is not ready. Specify Original to
get user-submitted template

My hypothesis is that the SynapseOwnerFile resource is attempting to
create a file however the bucket has not been created yet.
We add a DependsOn to make the file dependent on the bucket.