Fix S3 synapse bucket provisioning
Our Synapse S3 bucket template contains a macro (lambda).  When provisioning
the template thru the SC it failed to execute the lambda because it didn't have
permissions.  It would fail with the following (not helpful) message:

  Failed to execute transform 237179673806::S3Objects. Rollback requested by user.

Setting the permissions in the lambda itself does not work because the service
catalog role will override those permissions.  Fix this by giving the SC role
the full lambda permissions.