[SC-27] Provide SC users with connection info
Since SC users are given a dashboard of info after provisioning
SC products, the "provisioned products" list, the idea in this PR
is to use that dashboard to provide additional additional info
to users.  In this specific case we are providing info on how
to connected to a provisioned instance.